Type1/CFF font support; Acroform filling enhancements; basic object types enhancements; real number precision fixed (BUG 3409701).
git-svn-id: http://svn.code.sf.net/p/clown/code/trunk@48 68affd6e-c39b-4d6d-bb22-b3492c778056

+++ dotNET/pdfclown.lib/src/org/pdfclown/documents/contents/colorSpaces/Color.cs
@@ -49,8 +49,8 @@ namespace org.pdfclown.documents.contents.colorSpaces
       implementation of this method; currently this implementation focuses on
       device colors only.
     */
-    protected static float NormalizeComponent(
-      float value
+    protected static double NormalizeComponent(
+      double value
       )
     {
       if(value < 0)
